добавил валидацию,  добавил Эмитирование события для sidebar, поменял расположения полей

+++ src/components/telegram/Questions.vue
@@ -2,13 +2,34 @@
       <div>
   <TitleBlock :title="t('telegram.question')"/>
   <div class="col-12">
-        <Menubar :model="menu2" :key="active" style="height:36px;margin-top:-7px;margin-left:-14px;margin-right:-14px"></Menubar>
+        <ToolbarMenu :data="menu2" @search="search" :search="true" @filter="toggleFilter($event)" :filter="true" :filtered="filtered"/>
+
       </div>
-  <TabPanel>
-    <TreeTable ref="edutreetable" :scrollable="true" :scrollHeight="windowHeight + 'px'" class="p-treetable-sm"
-                 @node-select="onNodeSelect" :value="nodes" :lazy="true" :loading="loading"
-                 @nodeExpand="onExpand($event, true)" selectionMode="single" v-model:selectionKeys="selected">
-        <Column field="title_" :header="$t('telegram.question')" :expander="true"  style="min-width: 35rem;">
+      <div class="mb-3">
+      <OverlayPanel ref="filter">
+        <div v-for="text in menu_radio_options" :key="text" class="flex align-items-center">
+          <div class="field-radiobutton">
+            <RadioButton :value="text.value"/>
+            <label :for="text" class="ml-2">{{ text.text }}</label>
+          </div>
+        </div>
+        <div class="p-fluid">
+          <div class="field">
+            <br/>
+            <Button icon="pi pi-trash" class="p-button-outlined ml-1"
+                    :label="$t('common.clear')" @click="filterClick(false)"/>
+          </div>
+          <div class="field">
+            <Button icon="pi pi-search" :label="$t('common.search')" class="ml-1" @click="filterClick(true)"/>
+          </div>
+        </div>
+      </OverlayPanel>
+    </div>
+    <div class="card">
+                 <TreeTable class="p-treetable-sm" :value="nodes" :lazy="true" :loading="loading" @nodeExpand="onExpand"
+                     scrollHeight="flex" responsiveLayout="scroll" :resizableColumns="true" show-gridlines columnResizeMode="fit"
+                     :paginator="true" :rows="lazyParams.rows" :total-records="total" @page="onPage($event)">
+                 <Column field="title_" :header="$t('telegram.question')" expander style="min-width: 35rem;">
           <template #body="{node}">
               {{
                 node["title_" + locale]
@@ -35,23 +56,13 @@
               }}</span>
           </template>
         </Column>
-        <Column headerStyle="width: 10rem" >
-          <template #body="{node}">
-          <div class="inline-flex">
-          <Button class="p-button-text p-button-warning p-1 mr-2" @click="update(node)">
-            <i class="fa-solid fa-pencil fa-xl"></i>
-          </Button>
-
-          <Button class="p-button-text p-button-danger p-1 mr-2"  @click="deleteValue(node)">
-  <i class="fa-solid fa-trash-can fa-xl"></i>
-</Button>
-
-          </div>
+      <Column headerStyle="width: 10rem">
+        <template #body="{node}">
+          <ActionButton :show-label="true" :items="initItems" @toggle="toggle(node)"/>
         </template>
       </Column>
-
       </TreeTable>
-  </TabPanel>
+    </div>
   </div>
           <Sidebar v-model:visible="isView.node" position="right" class="p-sidebar-lg" style="overflow-y: scroll">
             <AddQuestions :modelValue=node :readonly="readonly"/>
@@ -60,45 +71,83 @@
 
 
 <script setup>
+import ToolbarMenu from "@/components/ToolbarMenu.vue";
 import Column from 'primevue/column';
 import TreeTable from 'primevue/treetable';
-import {ref, computed, onMounted} from 'vue';
+import {inject,ref, computed, onMounted} from 'vue';
 import {useI18n} from "vue-i18n";
-import {findRole} from "@/config/config";
 import {useToast} from "primevue/usetoast";
-import {createLogger, useStore} from "vuex";
-import {useRouter} from "vue-router";
 import AddQuestions from './AddQuetions.vue'
 import {QuestionService} from "@/service/question.service";
 
-const isView = ref({
-    node: false
-  })
-      // courseDialog: false,
-const lazyParams = {
+
+  const lazyParams = {
         page: 0,
         rows: 10,
         searchText: null,
+        parent_id: null,
       }
-
+const filtered = ref(false)
+const filter = ref()
 const node = ref (null);
+const emitter = inject("emitter");
 const nodes = ref([]);
 const questionService = new QuestionService
-const academicDegree = ref(null)
 const toast = useToast()
 const data = ref([]);
 const loading = ref(false);
 const {t, locale} = useI18n()
+const initItems = computed(() => {
+  return [
+  {
+      label: t('common.add'),
+      icon: 'fa-solid fa-plus',
+      command: () => {
+        createQuestion(node.value, node.value.id)
+            }
+    },
+    {
+      label: t('common.edit'),
+      icon: 'fa-solid fa-pen',
+      command: () => {
+        update(node.value)
+            }
+    },
+    {
+      label: t('common.delete'),
+      icon: 'fa-solid fa-trash',
+      command: () => {
+        deleteValue(node.value);  
+            }
+    },
+
+  ];
+})
+
+const isView = ref({
+    node: false
+  })
+
+const total = ref(0)
+const toggleFilter = (event) => {
+  filter.value.toggle(event)
+}
+const filterClick = (event) => {
+  filtered.value = event
+}
 const deleteValue =(node)=> {
+  if (!node) {
+    console.error('Node is undefined');
+    return;
+  }
       const req = {
         id: node.id,
     }
-
-    console.log(req)
     loading.value = true
     questionService.deleteQuestion(req).then(_ => {
         loading.value = false
         toast.add({severity: "success", summary: t('common.success'), life: 3000});
+        emitter.emit('node', true);
         getQuestions()
     }).catch(error => {
         toast.add({severity: 'error', summary: t('common.error'), life: 3000})
@@ -106,34 +155,22 @@ const deleteValue =(node)=> {
     })
   }
 
-
-// const update = () => {
-//   if (node.value != null) {
-//     isView.value.node = true
-//   }
-// }
-
 const update = (selectedNode) => {
   if (selectedNode != null) {
     node.value = selectedNode;
     isView.value.node = true;
   }
 }
-  
-const menu2= ref([
-        {
-          label: t("common.add"),
-          icon: "pi pi-fw pi-plus",
-          command: () => {
-            createQuestion()
-          },
-        },
-    ])
- const createQuestion=() => {
-      node.value = {}
-      node.value.userID = props.userID
-      isView.value.node = true
-  }
+
+
+
+const createQuestion = (parentNode = null) => {
+  node.value = {};
+  node.value.userID = props.userID;
+  node.value. parent_id = parentNode ? parentNode.id : null;
+  isView.value.node = true;
+}; 
+
   const props = defineProps({
     userID: {
       type: Number,
@@ -152,29 +189,75 @@ const menu2= ref([
       default: ''
     }
   });
+  
+  const toggle = (data) => {
+    node.value = data
+}
 
-  const getQuestions = () => {
+
+  const getQuestions = (node) => {
     const req = {
       page: lazyParams.page,
-      rows: lazyParams.rows
+      rows: lazyParams.rows,
+      parent_id: lazyParams.parent_id
     }
 
     loading.value = true
 
     questionService.getQuestions(req).then(res => {
-      nodes.value = res.data.questions
-      total.value = res.data.total
+      if (node == null) {
+        nodes.value = res.data.questions
+        total.value = res.data.total
+      } else {
+        console.log('res.data.questions', res.data.questions)
+        console.log('res.data.questions', res.data.questions)
+        node.value.children = res.data.questions
+      }
       loading.value = false
+
     }).catch(err => {
       loading.value=false
       toast.add({severity: 'error', summary: t('common.error'), life: 3000})
     })
 
   }
+  const menu2= computed(() => {
+  return [
+ {
+    label: t("common.add"),
+    icon: "pi pi-fw pi-plus",
+    //disabled: () => props.readonly,
+    command: () => {
+    createQuestion()
+          },
+        },
+      ]
+    })
+
+
+const onExpand = (data) => {
+  console.log(data.id)
+  lazyParams.parent_id = Number(data.id)
+  lazyParams.is_child = true
+  node.value = data
+  getQuestions(node)
+}
+
+const onPage = (event) => {
+    lazyParams.page = event.page + 1;
+    lazyParams.rows = event.rows
+    getQuestions(null);
+  }
+
   onMounted(() => {
     getQuestions()
+
+    emitter.on('node', (data) => {
+          if (data === true) {
+            isView.value.node = false
+            getQuestions()
+          }
+      });
   })
 
 </script>
-
-
